+++ .gitignore
@@ -5,3 +5,4 @@
 .kotlin/
 build/
 out/
+/pipeline/
