+++ formver.compiler-plugin/core/src/org/jetbrains/kotlin/formver/core/embeddings/callables/FullySpecialKotlinFunction.kt
@@ -178,7 +178,7 @@ object SpecialKotlinFunctions {
             args.map { Assert(it) }.toBlock()
         }
 
-        val forAllCallableType = buildFunctionPretype {
+        val quantifierCallableType = buildFunctionPretype {
             withParam {
                 function {
                     withDispatchReceiver {
@@ -199,7 +199,7 @@ object SpecialKotlinFunctions {
             }
         }
 
-        addFunction(forAllCallableType, SpecialPackages.formver, name = "forAll") { args, ctx ->
+        addFunction(quantifierCallableType, SpecialPackages.formver, name = "forAll") { args, ctx ->
             val arg = args.first()
             val lambda = arg.ignoringMetaNodes() as? LambdaExp ?: throw SnaktInternalException(
                 null,
@@ -213,7 +213,7 @@ object SpecialKotlinFunctions {
             ctx.insertForAllFunctionCall(param.symbol, body)
         }
 
-        addFunction(forAllCallableType, SpecialPackages.formver, name = "exists") { args, ctx ->
+        addFunction(quantifierCallableType, SpecialPackages.formver, name = "exists") { args, ctx ->
             val arg = args.first()
             val lambda = arg.ignoringMetaNodes() as? LambdaExp ?: throw SnaktInternalException(
                 null,
