+++ formver.compiler-plugin/testData/diagnostics/verification/user_invariants/exists.fir.diag.txt
@@ -1,51 +1,11 @@
-/exists.kt: info: Generated Viper text for maxCharacter:
-method maxCharacter(s: Ref) returns (v_ret_0: Ref)
+/exists.kt: info: Generated Viper text for simpleExists:
+method simpleExists(s: Ref) returns (v_ret_0: Ref)
   requires isSubtype(typeOf(s), stringType())
   requires |stringFromRef(s)| > 0
-  ensures isSubtype(typeOf(v_ret_0), charType())
-  ensures (forall anon_builtin_0: Int ::0 <= anon_builtin_0 &&
-      anon_builtin_0 < |stringFromRef(s)| ==>
-      stringFromRef(s)[anon_builtin_0] <= charFromRef(v_ret_0)) &&
-    (exists anon_builtin_1: Int :: 0 <= anon_builtin_1 &&
-      anon_builtin_1 < |stringFromRef(s)| &&
-      stringFromRef(s)[anon_builtin_1] == charFromRef(v_ret_0))
+  ensures isSubtype(typeOf(v_ret_0), intType())
+  ensures (exists anon: Int :: 0 <= anon && anon < |stringFromRef(s)|)
 {
-  var m: Ref
-  var i: Ref
-  var anon_0: Ref
-  m := stringGet(s, intToRef(0))
-  i := intToRef(1)
-  label cont
-    invariant isSubtype(typeOf(m), charType())
-    invariant isSubtype(typeOf(i), intType())
-    invariant isSubtype(typeOf(s), stringType())
-    invariant 1 <= intFromRef(i) && intFromRef(i) <= |stringFromRef(s)|
-    invariant (forall anon_builtin_2: Int ::0 <= anon_builtin_2 &&
-        anon_builtin_2 < intFromRef(i) ==>
-        stringFromRef(s)[anon_builtin_2] <= charFromRef(m))
-    invariant (exists anon_builtin_3: Int :: 0 <= anon_builtin_3 &&
-        anon_builtin_3 < intFromRef(i) &&
-        stringFromRef(s)[anon_builtin_3] == charFromRef(m))
-  anon_0 := ltInts(i, stringLength(s))
-  if (boolFromRef(anon_0)) {
-    if (stringFromRef(s)[intFromRef(i)] > charFromRef(m)) {
-      m := stringGet(s, i)
-    }
-    i := plusInts(i, intToRef(1))
-    goto cont
-  }
-  label break
-  assert isSubtype(typeOf(m), charType())
-  assert isSubtype(typeOf(i), intType())
-  assert isSubtype(typeOf(s), stringType())
-  assert 1 <= intFromRef(i) && intFromRef(i) <= |stringFromRef(s)|
-  assert (forall anon_builtin_2: Int ::0 <= anon_builtin_2 &&
-      anon_builtin_2 < intFromRef(i) ==>
-      stringFromRef(s)[anon_builtin_2] <= charFromRef(m))
-  assert (exists anon_builtin_3: Int :: 0 <= anon_builtin_3 &&
-      anon_builtin_3 < intFromRef(i) &&
-      stringFromRef(s)[anon_builtin_3] == charFromRef(m))
-  v_ret_0 := m
+  v_ret_0 := intToRef(0)
   goto lbl_ret_0
   label lbl_ret_0
 }
