--- formver.compiler-plugin/testData/diagnostics/verification/user_invariants/exists_char_bound.fir.diag.txt
@@ -1,19 +1,0 @@
-/exists_char_bound.kt: info: Generated Viper text for everyCharAtLeastNull:
-method everyCharAtLeastNull() returns (v_ret_0: Ref)
-  ensures isSubtype(typeOf(v_ret_0), intType())
-  ensures (forall anon: Int ::0 <= anon && anon < 65536 ==> 0 <= anon)
-{
-  v_ret_0 := intToRef(0)
-  goto lbl_ret_0
-  label lbl_ret_0
-}
-
-/exists_char_bound.kt: info: Generated Viper text for someCharIsNull:
-method someCharIsNull() returns (v_ret_0: Ref)
-  ensures isSubtype(typeOf(v_ret_0), intType())
-  ensures (exists anon: Int :: 0 <= anon && anon < 65536 && anon == 0)
-{
-  v_ret_0 := intToRef(0)
-  goto lbl_ret_0
-  label lbl_ret_0
-}
