--- formver.compiler-plugin/testData/diagnostics/verification/user_invariants/exists_char_bound.viper.diag.txt
@@ -1,1 +1,0 @@
-/exists_char_bound.kt: warning: Viper verification error: Postcondition of someCharIsNull might not hold. Assertion (exists anon: Int :: 0 <= anon && anon < 65536 && anon == 0) might not hold.
