--- pipeline/meta/f6-agent-ratings.md
@@ -1,29 +1,0 @@
-# F6 (`exists<T>`) — Agent Ratings (Step 9 Meta-Review)
-
-Impact = how much the agent moved the run forward. Quality = how well it did
-its job. Both 1–5. Ratings are of the agents' conduct in *this* run, judged
-against the run log and the on-disk artifacts (feedback doc, review, Step 6
-comparison, solver reports, implementation diff). Process only — no judgment of
-the compiler code itself.
-
-| agent | step | model | impact | quality | notes |
-| --- | --- | --- | :---: | :---: | --- |
-| orchestrator | 0 Planner | opus | 3 | 3 | Sound pick: `exists<T>` is atomic, correctly not decomposed. But the same orchestrator later trusted a spurious crash signal (spawned a duplicate) and initially tolerated a push to the examples-repo `main` — orchestration quality drags the score down. |
-| freya | 1 Fork | sonnet | 2 | 4 | Mechanical branch + pipeline docs; clean and uneventful. Low impact by nature, no friction introduced. |
-| wynne | 2 Implement | opus | 5 | 5 | The whole feature end-to-end, build PASS, publishToMavenLocal. Surfaced the Z3/`--max-workers` env facts and the "bare exists needs a witness invariant" limitation up front. Core deliverable of the run. |
-| gale | 2.5 Scaffold | opus | 4 | 4 | Built the examples harness every solver consumed; found the `all_targets` requirement, the array-index-in-pure-context wall, and the warnings-not-errors gotcha. High-leverage enabling infra. |
-| genevieve | 3 Solver A | opus | 5 | 5 | TwoEq VERIFIED; deep rough-edge findings (Char+Int uninterpreted, element arithmetic, unbounded Char quantifier, const-val scope, crash on contradictory postcondition). Isolated the crash and flagged the clean-build grep trap. |
-| nestor | 3 Solver B | sonnet | 5 | 5 | MaxElim + Downsampling core VERIFIED; Relaxed Prefix partial. Bisected the single cleanest crash repro (double String-index under one `exists`) that directly drove the fix, and pinned the split-point instantiation gap. Sonnet outperforming its tier. |
-| yelena | 3 Solver C | haiku | 2 | 1 | Both problems blocked; over-generalized to "nested quantifiers fail" — a misdiagnosis A and B refute on the *same* problems — then deleted its solution files, destroying the evidence. Net-negative diagnosis; the only redeeming impact is that a solver spuriously bouncing off verifiable problems became the headline case for prioritizing the crash fix. |
-| mateusz | 4 Synthesize | opus | 5 | 5 | Consolidated three reports, explicitly refuted C's misdiagnosis, pinpointed the crash to `Info.fromSilver` `else -> TODO` at `Info.kt:25`, and ranked the in-scope fixes. This document drove the debug step precisely. |
-| fyodor | 5 Debug | opus | 5 | 4 | Landed `547d1b04`: Info→NoInfo crash fix, `[0,65536)` Char bound, 2 regression tests, republish. Deferred fix #3 (split-point) with sound rationale. Docked one point: it genuinely stalled uncommitted and needed a manual operator resume to finish. |
-| calliope | 5 Debug (dup) | opus | 1 | 4 | Net-zero impact — a wasted opus build spawned on a false crash signal. But its conduct was exemplary for a duplicate: it independently converged on the identical fix, *detected the collision itself*, and stood down before commit/push. The waste is the orchestrator's fault, not calliope's. |
-| rafael | 6 Comparator | opus | 4 | 3 | Verdict ACCEPT, well-earned: clean-rebuilt and re-verified TwoEq/MaxElim/Downsampling-core (0 errors, no regression) and confirmed the crash→diagnostic end-to-end, even checking the mavenLocal jar's constant pool for provenance — strong rigor. Quality docked hard: it first pushed the deliverable to the examples repo's `main`, forcing a reset and a move to PR #1. |
-| marisol | 6.5 Clean | sonnet | 2 | 4 | Comment-only TODO strip; correctly kept the deliberate deferral docs. Good judgment, low impact. Never re-published — see the pipeline-feedback note on the accepted head being unverified-as-built. |
-| sidra | 7 Gate | sonnet | 3 | 3 | PASS on iteration 1. But REVIEW-ONLY (no build, host memory-contended) — a gate whose criteria include "compiles / one problem passes end-to-end" that leans on prior evidence rather than building is a weaker gate. Reasonable under constraints, not a strong assurance. |
-| quirin | 8 Review | opus | 4 | 5 | `pipeline/reviews/f6.md`: thorough, specific, and honest about limitations (String-only modelling, Char/Int arithmetic, split-point non-instantiation, witness-carrying burden). Accurately grounded in the diff and solver evidence. |
-
-Model-tier note: the two opus solvers (A) and the sonnet solver (B) both
-delivered verified problems and precise findings; the haiku solver (C) produced
-net-negative value (wrong root cause + deleted evidence). Sonnet (B, marisol)
-performed at or above tier throughout.
