--- pipeline/runs/f6.md
@@ -1,43 +1,0 @@
-# Run: F6 — Existential Quantifiers
-
-## Feature
-
-F6 — Existential Quantifiers.
-
-## Target problems
-
-- 2011 MaxElim
-- 2011 TreeMax
-- 2011 TwoEq
-- 2015 Relaxed Prefix
-- 2017 Maximum-sum subarray
-- 2018 Colored Tiles
-- 2019 Cartesian Trees
-- 2021 Lexicographic Permutations
-- 2022 Downsampling
-- 2024 Smart Array Copy
-- 2025 Minimum Excludant
-
-## Examples repo
-
-`komiputer/snakt-verifythis` — holds the VerifyThis problem archive under `problems/<year>/`.
-
-## Build strategy
-
-SnaKt is built on this branch and published via `./gradlew publishToMavenLocal`.
-Example solutions consume that published artifact; nothing is built inside the
-SnaKt tree.
-
-## Debug loop
-
-Max 5 iterations.
-
-## Progress log
-
-- Step 1: fork branch feature/f6-exists created, pipeline docs added.
-- Step 2: exists<T> implemented; ./gradlew build PASS; publishToMavenLocal done.
-- Step 3: 3 solvers ran (A/B/C); Step 4: consolidated feedback in pipeline/f6-feedback.md.
-- Step 5 (debug iter 1): fixed #4.1 (Info.fromSilver crash -> graceful Viper verification error) and #4.2/#4.4 char-quantifier auto-bound (forAll<Char>/exists<Char> now emit `0 <= x && x < 65536`); added regression tests exists_double_index (crash-now-diagnostic) and exists_char_bound (Char bound); full test suite PASS (first complete build) and detekt PASS after fixing a MagicNumber; publishToMavenLocal PASS (0.1.0-SNAPSHOT). Deferred: split-point exists instantiation (§4.2 boundary case) — triggers cannot supply a ground witness (matches solver B's exhausted attempts); a witness-hint DSL is out of scope for F6. Note: repeated ~11-min full `build` runs were OOM-killed on the shared host; validated via component tasks (test PASS, detekt PASS, publish PASS).
-- Step 6: comparator verdict ACCEPT (debugged strictly better). Step 6.5: stripped TODO-style comments from F6 code (comment-only).
-- Step 7: gate PASS (iter 1). Step 8: final review at pipeline/reviews/f6.md.
-- Step 9: meta-review at pipeline/meta/f6-agent-ratings.md and f6-pipeline-feedback.md.
